cambios de alejandro agregados

+++ src/componentes/Recordatorios.jsx
@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import Modal from 'react-modal';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faTrash } from '@fortawesome/free-solid-svg-icons';
 import { addDoc, collection, deleteDoc, doc, getDocs, serverTimestamp, setDoc, query, where, onSnapshot } from 'firebase/firestore';
 import { db, auth } from '../firebase/firebase';
 import { toast } from 'react-toastify';
@@ -19,30 +21,25 @@ const Recordatorios = () => {
   });
   const [modalOpen, setModalOpen] = useState(false);
   const [recordatorioEliminar, setRecordatorioEliminar] = useState(null);
-  const [confirmarEliminar, setConfirmarEliminar] = useState(false);
+  const [isLoading, setIsLoading] = useState(true);
+
+ 
 
   useEffect(() => {
     const user = auth.currentUser;
-
-   console.log(user);
     if (user) {
       const unsubscribe = onSnapshot(
         query(collection(db, 'recordatorios'), where('userId', '==', user.uid)),
         (snapshot) => {
           const datos = snapshot.docs.map((doc) => ({ id: doc.id, ...doc.data() }));
           setRecordatorios(datos);
-         
+          setIsLoading(false);
         }
       );
       return () => unsubscribe();
     }
- 
   }, []);
 
- 
-
-
-  
   const handleAgregarRecordatorio = async () => {
     const user = auth.currentUser;
     if (user) {
@@ -57,7 +54,7 @@ const Recordatorios = () => {
       try {
         const docRef = await addDoc(collection(db, 'recordatorios'), recordatorio);
         setRecordatorios([...recordatorios, { id: docRef.id, ...recordatorio }]);
-        setNuevoRecordatorio({ descripcion: '', fecha: '' });
+        setNuevoRecordatorio({ descripcion: '', frecuencia: '', fecha: '' });
         closeModal();
         toast.success('Recordatorio agregado correctamente');
       } catch (error) {
@@ -69,15 +66,18 @@ const Recordatorios = () => {
 
   const handleEliminarRecordatorio = async () => {
     const id = recordatorioEliminar.id;
-    try {
-      await deleteDoc(doc(db, 'recordatorios', id));
-      const actualizados = recordatorios.filter((r) => r.id !== id);
-      setRecordatorios(actualizados);
-      closeModal();
-      toast.success('Recordatorio eliminado correctamente');
-    } catch (error) {
-      console.error('Error al eliminar el recordatorio:', error);
-      toast.error('Error al eliminar el recordatorio');
+    const confirmar = window.confirm('¿Estás seguro de que deseas eliminar este recordatorio?');
+    if (confirmar) {
+      try {
+        await deleteDoc(doc(db, 'recordatorios', id));
+        const actualizados = recordatorios.filter((r) => r.id !== id);
+        setRecordatorios(actualizados);
+        closeModal();
+        toast.success('Recordatorio eliminado correctamente');
+      } catch (error) {
+        console.error('Error al eliminar el recordatorio:', error);
+        toast.error('Error al eliminar el recordatorio');
+      }
     }
   };
 
@@ -88,45 +88,53 @@ const Recordatorios = () => {
   const closeModal = () => {
     setModalOpen(false);
     setRecordatorioEliminar(null);
-    setConfirmarEliminar(false);
+    
   };
 
   const handleConfirmarEliminar = (recordatorio) => {
     setRecordatorioEliminar(recordatorio);
-    setConfirmarEliminar(true);
-    setModalOpen(true);
+    handleEliminarRecordatorio();
   };
-  // notificacion()
+
   return (
     <div className="cont">
-      <h2 className="recordatorios-heading">Recordatorios</h2>
+      <h2 className="title">Módulo de Recordatorios</h2>
 
-      <button onClick={openModal} className="agregar-recordatorio-btn">Agregar Recordatorio</button>
+      <button onClick={openModal} className="addButton">Agregar Recordatorio</button>
 
-      <ul className="recordatorios-list">
-        {recordatorios.map((recordatorio) => (
-          <li key={recordatorio.id} className="recordatorio-item">
-            <span>{recordatorio.descripcion}</span>
+      {isLoading ? (
+      <div className="loading-cont">
+        <div className="loading-bar">
+          <div className="loading-progress"></div>
+        </div>
+        <div className="loadingtext">Cargando Recordatorios...</div>
+      </div>
+      ) : (
+        <>
+          {recordatorios.length === 0 ? (
+            <div className='no-gastos'>
+              <p>Aún no hay recordatorios añadidos.</p>
+            </div>
             
-            <button onClick={() => handleConfirmarEliminar(recordatorio)} className="recordatorio-item-btn">Eliminar</button>
-          </li>
-        ))}
-      </ul>
-        
+          ) : (
+            <ul className="recordatorios-list">
+              {recordatorios.map((recordatorio) => (
+                <li key={recordatorio.id} className="recordatorio-item">
+                  <span>{recordatorio.descripcion}</span>
+                  
+                  <FontAwesomeIcon icon={faTrash} onClick={() => handleConfirmarEliminar(recordatorio)} className="recordatorio-item-btn"/>
+                </li>
+              ))}
+            </ul>
+          )}
+        </>
+      )}
       <Messaging />
 
       <Modal isOpen={modalOpen} onRequestClose={closeModal} appElement={document.getElementById('root')} className="modal" overlayClassName="overlay">
-        {confirmarEliminar ? (
-          <>
-            <h3>Eliminar Recordatorio</h3>
-            <p>¿Estás seguro de que deseas eliminar este recordatorio?</p>
-            <button type="button" onClick={handleEliminarRecordatorio} className="eliminar-recordatorio-btn">Eliminar</button>
-            <button type="button" onClick={closeModal} className="eliminar-recordatorio-btn">Cancelar</button>
-          </>
-        ) : (
-          <>
-            <h3>Agregar Recordatorio</h3>
-            <form>
+      <>
+        <h3 className='modalTitle'>Agregar Recordatorio</h3>
+        <form>
               <input
                 type="text"
                 value={nuevoRecordatorio.descripcion}
@@ -146,9 +154,8 @@ const Recordatorios = () => {
               <button type="button" onClick={handleAgregarRecordatorio} className="agregar-recordatorio-btn">Agregar</button>
               <button type="button" onClick={closeModal} className="agregar-recordatorio-btn">Cancelar</button>
             </form>
-          </>
-        )}
-      </Modal>
+      </>
+    </Modal>
     </div>
   );
 };
